relocate `Footer`
